refactor: add a pattern to names of archives
