+++ src/database/dataSource.ts
@@ -3,7 +3,7 @@ import dotenv from "dotenv";
 
 dotenv.config();
 
-const connection = new DataSource({
+const AppDataSource = new DataSource({
   type: "postgres",
   port: +process.env.TYPEORM_PORT,
   username: process.env.TYPEORM_USERNAME,
@@ -18,13 +18,4 @@ const connection = new DataSource({
   ssl: true,
 });
 
-connection
-  .initialize()
-  .then(() => {
-    console.log(`Data Source has been initialized`);
-  })
-  .catch((err) => {
-    console.error(`Data Source initialization error`, err);
-  });
-
-export default connection;
+export default AppDataSource;
